Create menu load module
Used a global array to store menu item details and then used a for of
loop in a function to render all menu items on the page. The point of
this project is to set up separate modules so I will likely come back
to style nicely at the end or when updating portfolio later.

Also, the array should like exist in a class, but that was not the point
of this lesson.